* Performed basic optimizations for tool parts and their stats using simple materials * Fixed an issue where it would not generate models for SecondaryMaterials

+++ src/main/java/com/sigmundgranaas/forgero/core/tool/ForgeroTool.java
@@ -30,7 +30,7 @@ public interface ForgeroTool {
 
     int getDurability();
 
-    int getAttackDamage();
+    float getAttackDamage();
 
     float getAttackSpeed();
